+++ src/se/fredsberg/euler/series/SumSeries.java
@@ -2,14 +2,13 @@ package se.fredsberg.euler.series;
 
 import se.fredsberg.euler.function.Function;
 
-
-public class SumSeries extends SumOverFunctionSeries{
+public class SumSeries extends SumOverFunctionSeries {
 
     public SumSeries() {
-        super(new Function<Integer>() {
+        super(new Function<Long>() {
 
             @Override
-            public Integer calculate(Integer t) {
+            public Long calculate(Long t) {
                 return t;
             }
         });
